me olvidava de los constructores
ya esta

+++ Vector3/Vector3/Vector3.cpp
@@ -3,7 +3,7 @@
 
 Vector3::Vector3(){}
 
-Vector3::Vector3(int x, int y, int z) : x(x), y(y), z(z) {}
+Vector3::Vector3(float x, float y, float z) : x(x), y(y), z(z) {}
 
 Vector3::Vector3(Vector3 &vec3) : x(vec3.x), y(vec3.y), z(vec3.z) {}
